+++ src/phd_hunter/crawlers/__init__.py
@@ -3,6 +3,7 @@
 from .base import BaseCrawler, CacheEntry
 from .csrankings import CSRankingsCrawler
 from .arxiv_crawler import ArxivCrawler
+from .openalex_crawler import OpenAlexCrawler
 from .homepage_crawler import fetch_and_summarize_homepage, batch_fetch_homepages
 
 __all__ = [
@@ -10,6 +11,7 @@ __all__ = [
     "CacheEntry",
     "CSRankingsCrawler",
     "ArxivCrawler",
+    "OpenAlexCrawler",
     "fetch_and_summarize_homepage",
     "batch_fetch_homepages",
 ]